refactor(symbols): split 1575-LOC handlers.rs into 9 sub-modules
Pure relocation — no behavior changes. tools/symbols/handlers.rs (1,575
LOC) is decomposed into handlers/ with bm25, confidence, find_symbol,
follow_up, housekeeping, overview, path_args, ranked_context, mod.

Cross-module wiring:
- bm25 imports confidence_tier / suggested_follow_up / flatten_symbols /
  resolve_path_argument / insert_response_annotations via use super::.
- ranked_context imports sparse_symbol_hits_for_query +
  adapt_budget_to_context_window from bm25.
- Inline test modules (find_symbol_argument_tests, adapt_budget_tests)
  reference cross-module fns via super::super::<sub>::<fn>.

Visibility cost (pub(super) — minimal cross-boundary exposure):
  resolve_path_argument, insert_response_annotations (path_args)
  sparse_symbol_hits_for_query (bm25)
  suggested_follow_up (follow_up)
  confidence_tier (confidence)
adapt_budget_to_context_window keeps pub(crate) from the original file.

mod.rs preserves the external surface: bm25_symbol_search, find_symbol,
flatten_symbols, get_complexity, refresh_symbol_index,
search_symbols_fuzzy, get_symbols_overview, get_ranked_context — all
cross-crate callers (verifier_reports, context_reports, inline,
change_signature, etc.) require zero changes.

Verification:
- cargo check -p codelens-mcp --features http,semantic — PASS
- cargo test -p codelens-mcp --bin codelens-mcp --features http,semantic
  — 724/724 PASS in 63s
- cargo fmt --all — applied
