+++ crates/codelens-mcp/src/tools/symbols/handlers/housekeeping.rs
@@ -1,0 +1,176 @@
+use super::super::super::{
+    AppState, ToolResult, optional_bool, optional_string, optional_usize,
+    query_analysis::{RetrievalQueryAnalysis, analyze_retrieval_query},
+    required_string, success_meta,
+};
+use super::super::{
+    analyzer::{
+        annotate_ranked_context_provenance, compact_semantic_evidence, compact_sparse_evidence,
+        merge_semantic_ranked_entries, merge_sparse_ranked_entries, semantic_results_for_query,
+        semantic_scores_for_query,
+    },
+    formatter::{compact_symbol_bodies, count_branches},
+};
+use crate::error::CodeLensError;
+use crate::protocol::BackendKind;
+use crate::symbol_corpus::build_symbol_corpus;
+use crate::symbol_retrieval::{ScoredSymbol, search_symbols_bm25f, unique_query_terms};
+use codelens_engine::{SymbolInfo, SymbolKind, read_file, search_symbols_hybrid_with_semantic};
+use serde_json::{Value, json};
+
+use super::path_args::{insert_response_annotations, resolve_path_argument};
+
+pub fn refresh_symbol_index(state: &AppState, _arguments: &Value) -> ToolResult {
+    let stats = state.symbol_index().refresh_all()?;
+    state.graph_cache().invalidate();
+    #[cfg(feature = "semantic")]
+    let mut payload = json!(stats);
+    #[cfg(not(feature = "semantic"))]
+    let payload = json!(stats);
+    #[cfg(feature = "semantic")]
+    {
+        let project = state.project();
+        let guard = state.embedding_ref();
+        if let Some(engine) = guard.as_ref()
+            && engine.is_indexed()
+        {
+            match engine.ensure_index_fresh_for_project(&project) {
+                Ok(report) => {
+                    if let Some(map) = payload.as_object_mut() {
+                        map.insert("embedding_freshness".to_owned(), json!(report));
+                    }
+                }
+                Err(error) => {
+                    if let Some(map) = payload.as_object_mut() {
+                        map.insert(
+                            "embedding_freshness".to_owned(),
+                            json!({
+                                "status": "unavailable",
+                                "reason": error.to_string()
+                            }),
+                        );
+                    }
+                }
+            }
+        }
+    }
+    Ok((payload, success_meta(BackendKind::TreeSitter, 0.95)))
+}
+
+pub fn get_complexity(state: &AppState, arguments: &Value) -> ToolResult {
+    let path = required_string(arguments, "path")?;
+    let symbol_name = optional_string(arguments, "symbol_name");
+    let file_result = read_file(&state.project(), path, None, None)?;
+    let lines = file_result.content.lines().collect::<Vec<_>>();
+    let symbols = state.symbol_index().get_symbols_overview_cached(path, 2)?;
+
+    let functions = flatten_symbols(&symbols)
+        .into_iter()
+        .filter(|s| matches!(s.kind, SymbolKind::Function | SymbolKind::Method))
+        .filter(|s| symbol_name.is_none_or(|name| s.name == name))
+        .map(|s| {
+            let start = s.line.saturating_sub(1).min(lines.len());
+            let end = (s.line + 50).min(lines.len());
+            let branches = count_branches(&lines[start..end]);
+            json!({
+                "name": s.name,
+                "kind": s.kind.as_label(),
+                "file": s.file_path,
+                "line": s.line,
+                "branches": branches,
+                "complexity": 1 + branches
+            })
+        })
+        .collect::<Vec<_>>();
+
+    let results = if functions.is_empty() {
+        let branches = count_branches(&lines);
+        vec![json!({
+            "name": path,
+            "branches": branches,
+            "complexity": 1 + branches
+        })]
+    } else {
+        functions
+    };
+
+    let avg_complexity = if results.is_empty() {
+        0.0
+    } else {
+        results
+            .iter()
+            .filter_map(|e| e.get("complexity").and_then(|v| v.as_i64()))
+            .map(|v| v as f64)
+            .sum::<f64>()
+            / results.len() as f64
+    };
+
+    Ok((
+        json!({
+            "path": path,
+            "functions": results,
+            "count": results.len(),
+            "avg_complexity": avg_complexity
+        }),
+        success_meta(BackendKind::TreeSitter, 0.89),
+    ))
+}
+
+pub fn search_symbols_fuzzy(state: &AppState, arguments: &Value) -> ToolResult {
+    let query = required_string(arguments, "query")?;
+    let max_results = optional_usize(arguments, "max_results", 30);
+    let fuzzy_threshold = arguments
+        .get("fuzzy_threshold")
+        .and_then(|v| v.as_f64())
+        .unwrap_or(0.6);
+    let disable_semantic = optional_bool(arguments, "disable_semantic", false);
+    // Build semantic scores if embeddings are available (same pattern as get_ranked_context)
+    let semantic_scores = semantic_scores_for_query(state, query, 50, disable_semantic);
+
+    let sem_ref = if semantic_scores.is_empty() {
+        None
+    } else {
+        Some(&semantic_scores)
+    };
+
+    let backend = if sem_ref.is_some() {
+        BackendKind::Hybrid
+    } else {
+        BackendKind::Sqlite
+    };
+
+    let pagerank_scores = state.graph_cache().file_pagerank_scores(&state.project());
+    let pagerank_ref = if pagerank_scores.is_empty() {
+        None
+    } else {
+        Some(pagerank_scores.as_ref())
+    };
+
+    Ok(search_symbols_hybrid_with_semantic(
+        &state.project(),
+        query,
+        max_results,
+        fuzzy_threshold,
+        sem_ref,
+        pagerank_ref,
+    )
+    .map(|value| {
+        (
+            json!({ "results": value, "count": value.len() }),
+            success_meta(backend, 0.9),
+        )
+    })?)
+}
+
+// ── Helpers ──────────────────────────────────────────────────────────────
+
+pub fn flatten_symbols(symbols: &[SymbolInfo]) -> Vec<SymbolInfo> {
+    let mut flat = Vec::new();
+    let mut stack = symbols.to_vec();
+    while let Some(mut symbol) = stack.pop() {
+        let children = std::mem::take(&mut symbol.children);
+        flat.push(symbol);
+        stack.extend(children);
+    }
+    flat
+}
