+++ crates/codelens-mcp/src/tools/symbols/handlers/mod.rs
@@ -1,0 +1,16 @@
+mod bm25;
+mod confidence;
+mod find_symbol;
+mod follow_up;
+mod housekeeping;
+mod overview;
+mod path_args;
+mod ranked_context;
+
+pub use bm25::bm25_symbol_search;
+pub use find_symbol::find_symbol;
+pub use housekeeping::{
+    flatten_symbols, get_complexity, refresh_symbol_index, search_symbols_fuzzy,
+};
+pub use overview::get_symbols_overview;
+pub use ranked_context::get_ranked_context;
